# Conflicts:
#	frontend/src/main/scala/com/neo/sk/gypsy/front/gypsyClient/GameHolder.scala

+++ shared/src/main/scala/com/neo/sk/gypsy/shared/ptcl/Protocol.scala
@@ -48,13 +48,13 @@ object Protocol {
 
   case class PlayerRestart(id:String) extends GameMessage
 
-  case class UserDeadMessage(id:String,killerId:String,killerName:String,killNum:Int,score:Int,lifeTime:Long) extends GameMessage
+  case class UserDeadMessage(id:String,killerId:String,killerName:String,killNum:Int,score:Short,lifeTime:Long) extends GameMessage
 
   case class Wrap(ws:Array[Byte],isKillMsg:Boolean = false) extends WsMsgSource
 
   case class KillMessage(killerId:String,deadPlayer:Player) extends GameMessage
 
-  case class GameOverMessage(id:String,killNum:Int,score:Int,lifeTime:Long) extends GameMessage
+  case class GameOverMessage(id:String,killNum:Int,score:Short,lifeTime:Long) extends GameMessage
 
   case class MatchRoomError() extends GameMessage
 
@@ -93,7 +93,7 @@ object Protocol {
 
 //  case class TextInfo(msg:String) extends UserAction
 
-  case class MousePosition(id: String,clientX:Double,clientY:Double, override val frame:Long, override val serialNum:Int) extends UserAction with GameMessage
+  case class MousePosition(id: String,clientX:Short,clientY:Short, override val frame:Long, override val serialNum:Int) extends UserAction with GameMessage
 
   case class KeyCode(id: String,keyCode: Int, override val frame:Long,override val serialNum:Int) extends UserAction with GameMessage
 
@@ -129,11 +129,11 @@ object Protocol {
   case class UserJoinRoom(roomId:Long,playState:Player, override val frame:Long) extends GameEvent
   case class UserLeftRoom(userId:String,userName:String,ballId:Long,roomId:Long, override val frame:Long) extends GameEvent
   case class UserWsJoin(roomId:Long,userId:String,userName:String,ballId:Long, override val frame:Long, override val serialNum:Int) extends GameEvent //每次webSocket加入时候记，不记Play的具体状态
-  case class MouseMove(userId:String,direct:(Double,Double), override val frame:Long, override val serialNum:Int) extends GameEvent
+  case class MouseMove(userId:String,direct:(Short,Short), override val frame:Long, override val serialNum:Int) extends GameEvent
   case class KeyPress(userId:String,keyCode: Int, override val frame:Long, override val serialNum:Int) extends GameEvent
-  case class GenerateApples(apples:Map[Point, Int], override val frame:Long) extends GameEvent
+  case class GenerateApples(apples:Map[Point, Short], override val frame:Long) extends GameEvent
   case class GenerateVirus(virus: Map[Long,Virus], override val frame:Long) extends GameEvent with WsMsgSource
-  case class KillMsg(killerId:String,deadPlayer:Player,score:Int,lifeTime:Long, override val frame: Long) extends GameEvent
+  case class KillMsg(killerId:String,deadPlayer:Player,score:Short,lifeTime:Long, override val frame: Long) extends GameEvent
   case class CurrentRanks(currentRank: List[RankInfo]) extends GameEvent
   case class PongEvent(timestamp: Long)extends GameEvent
 
@@ -170,7 +170,5 @@ object Protocol {
                                   //     boundary: ,
                                   //     window:
                                 )
-
-
 
 }
